requirements.txt


chromadb>=0.5.0
streamlit
requests
python-dotenv

sentence-transformers
langchain
langchain-google-genai
tqdm



